removed runnerhead.as include in quarters.as
this just simplifies the code. my main motive for this is to be able to add hooks like onInit and onCommand into runnerhead. before this change it was not possible because the compiler thinks there is already a function with the same name and parameter present (since runnerhead is included at the top). this change will allow me to update the playerheads mod without having to make changes to the quarters script. this helps to keep the playerheads mod to be as compatible with other mods as much as possible. thanks for reading