[modified] rewritten and edited almost all strings in Russian locale
So now it's a high-quality translation. 
Some people made their translation a month ago, and they noticed that somebody else made his file an official locale. So they contacted me, said that my translation is awful (I agree with that, cause it was made in about an hour). So we collaborated and made a new better translation. I rewrote and edited almost all death tips in their translation and merged them with my file, also edited a lot of other things.