Add property that allows changing buy offset
In your shop blob add `this.set_bool("shop offset is buy offset", true);` to change the "shop location" to wherever the shop offset is, this way you can create a shop outside of the blob's radius. For example, setting this property to true in war base you can actually buy things in the workbench.