Makes bomb arrow damage consistant
As requested in bounty # 3 for July 2018.

Fixes bomb arrows acting strangely by making them explode immediatly on contact with enemies. This will deal 1.5 damage to the hit enemy.

This was tested on a server for a while in a number of scenarios, however it's possible the issue may still exist. If it does I'll be happy to fix it.