Fix divide-by-zero in VehicleGUI when max_charge_time is zero
Fixes divide-by-zero error in VehicleGUI.as and therefore allows non-default vehicles to use instant-fire ammo (that is, ammo with max_charge_time = 0).

When you add a custom vehicle without charge delay it crashes when trying to fire because of progressbar percent division. The original code worked for mounted bow because it was explicitly excluded.