[dev] apply suggested changes
Also, I rewrote the function that looks for respawn spots in CTF_GiveSpawnItems.as.
Instead of the spots looking for the player overlapping with them, it is now the player looking for spots overlapping with him.
Also, return; when all conditions are met - no need to look through the remaining overlapping blobs.