[dev] add coin reward, add collidewhenattached, fixes
Added "collideWhenAttached", which is also remembered and reverted like "collidable" when the tipped blob is picked up or the arrow dies.
"collideWhenAttached" is set to true while the Arrow is flying, so Foods can heal teammates.
Added coin reward for doing so.

Only attach non-attached blob to the arrow's tip.