[dev] some changes
Instead of assigning random team 2-6 in TDM, Dummy will instead spawn as team 255 and get a random color set via ReloadSprites(). This is the same behavior that Fishy uses when spawning in water.

Challenge.as will still set team to 1 for tutorial maps. Additionally, Left-facing will be set to true.

Dummy will now always be neutral team but have a random color in all game modes, unless the rules script overrides it.