[added] g_holiday_assets support + fixes
Fixed the birthday logic not applying properly due to the default
holiday being associated with the 0 value, which is what the rules
variable initializes to.

Other cleanups to the logic also made.

Added the script side plumbing for g_holiday_assets support. This allows
clients to disable most of the asset swaps when disabled. The engine
takes care of disabling the asset swaps for the main menu.
This disables cosmetics such as sprites, music and sounds that would
otherwise be selected when the relevant CFileMatcher overlay is enabled.
This includes changes to holiday logic, to disable snow during christmas
and to restore the default skygradient style. Since both Christmas and
Halloween scripts override the daytime cycle, and that the day time is
forcefully synced by the server, we work around this by adding a
skygradient_dayonly texture.