[new] Enable active sync on a few blobs
The TLDR is that the blob is synced whilst the box2d body is active (you
can check this with g_debug 1). You only want to use this on blobs where
knowing where the blob position is very important (e.g. players,
projectiles, explosives etc). The trade off is that this will consume
more bandwidth.

Disabling this/not setting it will use the default sync method, where
every tick, we check to see if our blob position has moved in any
direction far enough for the sync to be worth it.